Begun implementing the ShapeRenderer

+++ OpenGL/src/engine/Shader.h
@@ -24,7 +24,7 @@ public:
 	~Shader();
 	
 	void setUniform1i(const std::string& name, int value);
-	void setUniform1iv(const std::string& name, unsigned int count, const int* value);
+	void setUniform1iv(const std::string& name, size_t count, const int* value);
 	void setUniform4f(const std::string& name, float v1, float v2, float v3, float v4);
 	void setUniformMat4(const std::string& name, const Mat4& matrix);
